@@ -1,3 +1,3 @@
-module github.com/hashicorp/golang-lru
+module github.com/hashicorp/golang-lru/v2
 
-go 1.12
+go 1.18
